refactor: split rest_api/v1 views by domain ownership and package layout
Moves the five Admin-Console-shaped endpoints (AdminConsoleOrgsAPIView,
ScopesAPIView, TeamMembersAPIView, TeamMemberAssignmentsAPIView,
AssignmentsAPIView) into a new admin_console/ subpackage, and
WaffleFlagStatesAPIView into a new course_authoring/ subpackage, per
the ownership and layout rules being recorded in ADR 0016 and ADR 0017.
PermissionValidationMeView, RoleUserAPIView, RoleListView, and
UserValidationAPIView remain in views.py as generic Authorization
endpoints, with no reference to any specific consumer's packaging;
that reasoning lives in admin_console/course_authoring's own docstrings
and in ADR 0017. No URLs change; tests move alongside their views.
