+++ openedx_authz/rest_api/v1/course_authoring/__init__.py
@@ -1,0 +1,7 @@
+"""REST API views exposing Course Authoring data.
+
+This package holds the one documented exception to openedx-authz's domain
+boundary. This is decision item 3 in ADR 0017 (rest-api-package-layout-convention),
+a subpackage named after the domain that actually owns this data. See ADR 0015
+for why the exception exists and ADR 0016 for the domain-ownership determination.
+"""
